4.2 Tablice powtórzenie. Zadanie: Losowanie wartości i iterowanie po tablicy.

+++ kodilla-collections/src/main/java/com/kodilla/collections/interfaces/homework/Volvo.java
@@ -6,7 +6,7 @@ public class Volvo implements Car {
 
     private double speed;
 
-    public Volvo (int speed){
+    public Volvo (double speed){
         this.speed = speed;
     }
